add helper lemma for proof generation of CFG optimizations
This helper lemma is used in cases when the loop head set representation
is different in a proof goal and in a lemma. Previously, the simplifier
was used on the proof goal, but the result does not always match the
representation of the lemma that is then applied. Now, the helper lemma
is applied instead.